requirements.txt
aiogram==3.7.0
aiosqlite==0.19.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
APScheduler==3.10.4
openai==1.3.0
